Add testnet3 support: lax DER parsing, BIP16 activation heights, P2P improvements
- Lax DER signature parser for pre-BIP66 transactions (verify_ecdsa now
  uses _parse_sig_der_lax + secp256k1_ecdsa_signature_parse_compact,
  matching Bitcoin Core's CPubKey::Verify approach)
- Per-network BIP16 (P2SH) activation heights: mainnet=173805,
  testnet3=514, testnet4/signet=1, regtest=0
- Correct testnet3 BIP activation heights (BIP34=21111, BIP65=581885,
  BIP66=330776, CSV=770112, SegWit=834624, Taproot=2032291)
- Fix mainnet pow_limit (4 leading zero bytes, not 3)
- Fix i64 timespan arithmetic in difficulty retarget (prevents unsigned
  overflow when testnet3 timestamps go backwards)
- 5-second TCP connect timeout with non-blocking socket + poll
- Gradual peer connection (connect first peer synchronously, rest in
  event loop) with non-blocking recv during peer connection phase
- Block progress logging every 1000 blocks during IBD
- Regression test: test_testnet3_26860_p2pkh (non-strict DER signature)

+++ consensus/pow.odin
@@ -129,11 +129,13 @@ calculate_next_work_required :: proc(
 		return current_bits
 	}
 
-	actual_timespan := current_time - last_retarget_time
+	// Use signed arithmetic — on testnet3, timestamps can go backwards
+	// due to the 20-minute min-difficulty rule (miners set future timestamps).
+	actual_timespan := i64(current_time) - i64(last_retarget_time)
 
 	// Clamp to [target_timespan/4, target_timespan*4]
-	min_timespan := params.target_timespan / 4
-	max_timespan := params.target_timespan * 4
+	min_timespan := i64(params.target_timespan) / 4
+	max_timespan := i64(params.target_timespan) * 4
 	if actual_timespan < min_timespan {
 		actual_timespan = min_timespan
 	}
